Music/Allianza Blockchain interoperabilidade funciona 100%/Allianza Blockchain/requirements_uec.txt
# requirements_uec.txt - ATUALIZADO
web3>=6.0.0
bitcoinlib>=0.12.0
requests>=2.28.0
base58>=2.1.1
bech32>=1.2.0
ecdsa>=0.18.0
cryptography>=41.0.0
flask>=2.3.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
# Adicionar no requirements_uec.txt:
web3>=6.0.0
bitcoinlib>=0.12.0
python-dotenv>=1.0.0